+++ packages/ragbits-agents/CHANGELOG.md
@@ -1,6 +1,18 @@
 # CHANGELOG
 
 ## Unreleased
+- TodoOrchestrator & Todo list for agent (#823)
+
+- Support wrapping downstream agents as tools (#818)
+- Add syntax sugar allowing easier Agents definition (#820)
+- Add post-processors (#821)
+- Support streaming from downstream agents (#812)
+
+- Support wrapping downstream agents as tools (#818)
+- Add syntax sugar allowing easier Agents definition (#820)
+- Add post-processors (#821)
+- Support streaming from downstream agents (#812)
+- Add supervisor post-processor (#830)
 
 ## 1.3.0 (2025-09-11)
 ### Changed